fix: fix coderabbit issues
