+++ src/test/java/org/storkforge/barkr/web/controller/WebControllerTest.java
@@ -7,17 +7,9 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
-import org.springframework.http.HttpHeaders;
-import org.springframework.mock.web.MockMultipartFile;
-import org.springframework.test.context.bean.override.mockito.MockitoBean;
-import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
-import org.storkforge.barkr.domain.AccountService;
-import org.storkforge.barkr.domain.DogFactService;
-import org.storkforge.barkr.domain.PostService;
 import org.storkforge.barkr.domain.entity.Account;
 import org.storkforge.barkr.domain.entity.Post;
 import org.storkforge.barkr.infrastructure.persistence.AccountRepository;
@@ -26,19 +18,12 @@ import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 import org.testcontainers.utility.DockerImageName;
-import org.springframework.core.io.ClassPathResource;
 
 import java.io.IOException;
-import java.io.InputStream;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.mockito.Mockito.doNothing;
-import static org.mockito.Mockito.when;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.multipart;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @AutoConfigureMockMvc
@@ -151,7 +136,8 @@ class WebControllerTest {
       HtmlTextArea contentInput = form.getTextAreaByName("content");
       HtmlButton submitButton = (HtmlButton) form.getElementsByTagName("button").getFirst();
 
-      contentInput.setText("THIS STRING IS 256 CHARACTERS xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
+      String tooLongContent = "A".repeat(256);
+      contentInput.setText(tooLongContent);
 
       HtmlPage resultPage = submitButton.click();
 
@@ -200,95 +186,5 @@ class WebControllerTest {
 
       assertThat(((HtmlDivision) page.getFirstByXPath("//div[@class='error-message']")).getTextContent()).contains("User 'mockAccount' not found");
     }
-  }
-}
-
-@WebMvcTest(WebController.class)
-class WebControllerImageRouteTest {
-  @Autowired
-  private MockMvc mockMvc;
-
-  @MockitoBean
-  private AccountService accountService;
-
-  @MockitoBean
-  private PostService postService;
-
-  @MockitoBean
-  private DogFactService dogFactService;
-
-  @Test
-  @DisplayName("Returns account image if present")
-  void returnsAccountImageIfPresent() throws Exception {
-    byte[] image = "test-image".getBytes();
-    when(accountService.getAccountImage(1L)).thenReturn(image);
-
-    mockMvc.perform(get("/account/1/image"))
-            .andExpect(status().isOk())
-            .andExpect(header().string(HttpHeaders.CONTENT_TYPE, "application/octet-stream"))
-            .andExpect(content().bytes(image));
-  }
-
-  @Test
-  @DisplayName("Returns fallback image if account image is null")
-  void returnsFallbackImageIfNull() throws Exception {
-    byte[] fallback;
-    try (InputStream is = new ClassPathResource("static/images/logo/BarkrNoText.png").getInputStream()) {
-      fallback = is.readAllBytes();
-    }
-
-    when(accountService.getAccountImage(1L)).thenReturn(null);
-
-    mockMvc.perform(get("/account/1/image"))
-            .andExpect(status().isOk())
-            .andExpect(header().string(HttpHeaders.CONTENT_TYPE, "application/octet-stream"))
-            .andExpect(content().bytes(fallback));
-  }
-
-  @Test
-  @DisplayName("Redirects with error if file is empty")
-  void fileIsEmpty() throws Exception {
-    MockMultipartFile emptyFile = new MockMultipartFile("file", "image.png", "image/png", new byte[0]);
-
-    mockMvc.perform(multipart("/account/1/upload").file(emptyFile))
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl("/"))
-            .andExpect(flash().attribute("error", "File is empty"));
-  }
-
-  @Test
-  @DisplayName("Redirects with error if file is not an image")
-  void fileIsNotImage() throws Exception {
-    MockMultipartFile file = new MockMultipartFile("file", "file.txt", "text/plain", "some-text".getBytes());
-
-    mockMvc.perform(multipart("/account/1/upload").file(file))
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl("/"))
-            .andExpect(flash().attribute("error", "File is not an image"));
-  }
-
-  @Test
-  @DisplayName("Redirects with error if file is too large")
-  void fileIsTooLarge() throws Exception {
-    byte[] large = new byte[6 * 1024 * 1024]; // 6MB
-    MockMultipartFile file = new MockMultipartFile("file", "big.png", "image/png", large);
-
-    mockMvc.perform(multipart("/account/1/upload").file(file))
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl("/"))
-            .andExpect(flash().attribute("error", "File is too large"));
-  }
-
-  @Test
-  @DisplayName("Redirects to root after successful image upload")
-  void uploadSuccess() throws Exception {
-    byte[] img = "image-content".getBytes();
-    MockMultipartFile file = new MockMultipartFile("file", "pic.png", "image/png", img);
-
-    doNothing().when(accountService).updateImage(1L, img);
-
-    mockMvc.perform(multipart("/account/1/upload").file(file))
-            .andExpect(status().is3xxRedirection())
-            .andExpect(redirectedUrl("/"));
   }
 }
